# specs/005-contracts-module/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: FlyAccounts
  description: 'FlyAccounts API 6.0.0: global Clients CRUD; server-side search/pagination on clients,
    contracts, people, roles; contracts require clientId on create.'
  version: 6.0.0
servers:
- url: /v1
  description: Versioned API prefix
...
paths:
...
  /people:
    get:
      operationId: listPeople
      summary: Signed-in people and unused invites
      description: 'Requires manage_users in the caller''s combined permissions. Returns a union of Users
        and active Invites. Unused invites have personType invite and status invited. Signed-in people
        have personType user and status pending or active.

        '
      tags:
      - Settings
      security:
      - bearerAuth: []
      responses:
        '200':
          description: People and unused invites with roles.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PeopleResponse'
        '401':
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Combined permissions do not include the required manage_* permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      parameters:
      - &id001
        $ref: '#/components/parameters/ListSearch'
      - &id002
        $ref: '#/components/parameters/ListPage'
      - &id003
        $ref: '#/components/parameters/ListPageSize'
...
  /people/{userId}:
    delete:
      operationId: removePerson
      summary: Remove a signed-in person
      description: 'Requires manage_users. Removes a listed User and their role assignments. Refresh tokens
        for that person are revoked. Does not prevent a later organization-based sign-in by the same account.
        Refused when the change would leave zero signed-in people whose combined permissions include manage_users.
        Unused invites are cancelled with DELETE /people/invites/{inviteId}, not this operation.

        '
      tags:
      - Settings
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: Person removed.
        '401':
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Combined permissions do not include the required manage_* permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Person not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Would remove the last access-administration grant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: last_admin_required
                message: At least one person with manage_users must remain.
  /people/{userId}/roles:
    post:
      operationId: assignRoles
      summary: Assign one or more existing roles
      description: 'Adds the roles without replacing roles the person already has. Duplicate assignment
        of any role is refused. Requires manage_users.

        '
      tags:
      - Settings
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRoleRequest'
      responses:
        '200':
          description: Person after the assignment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '401':
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Combined permissions do not include the required manage_* permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Person or role not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The person already has this role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: duplicate_assignment
                message: That role is already assigned.
  /people/{userId}/roles/{roleId}:
    delete:
      operationId: revokeRole
      summary: Revoke one assigned role
      description: 'Removes one assignment. Remaining roles stay. If none remain, the person has no roles
        (pending access on their next use). Refused when the change would leave zero signed-in people
        whose combined permissions include manage_users. Unused invites do not count as administrators.

        '
      tags:
      - Settings
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/UserId'
      - $ref: '#/components/parameters/RoleId'
      responses:
        '200':
          description: Person after the revoke.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '401':
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Combined permissions do not include the required manage_* permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Person, role, or assignment not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Would remove the last access-administration grant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: last_admin_required
                message: At least one person with manage_users must remain.
  /people/invites:
    post:
      operationId: createInvite
      summary: Invite a person by organizational work email
      description: 'Requires manage_users. Creates an unused invite with zero or more existing roles.
        The invited person cannot use the application until they sign in. Duplicate active invite of the
        same email is refused. Invite of an email that already belongs to a listed User is refused. Sending
        email is not part of this operation.

        '
      tags:
      - Settings
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInviteRequest'
      responses:
        '201':
          description: Invite recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '401':
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Combined permissions do not include the required manage_* permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: A listed roleId was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already invited or already belongs to a listed person.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                duplicateInvite:
                  value:
                    code: duplicate_invite
                    message: That email is already invited.
                alreadyPresent:
                  value:
                    code: already_present
                    message: That person is already listed.
  /people/invites/{inviteId}:
    delete:
      operationId: cancelInvite
      summary: Cancel an unused invite
      description: 'Requires manage_users. Removes an unused invite from the list. Does not prevent a
        later organization-based sign-in by that email. Consumed invites cannot be cancelled.

        '
      tags:
      - Settings
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/InviteId'
      responses:
        '204':
          description: Invite cancelled.
        '401':
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Combined permissions do not include the required manage_* permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Active unused invite not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /people/invites/{inviteId}/roles:
    post:
      operationId: assignInviteRoles
      summary: Assign one or more existing roles to an unused invite
      description: 'Adds the roles without replacing roles the invite already has. Duplicate assignment
        of any role is refused. Requires manage_users.

        '
      tags:
      - Settings
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/InviteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRoleRequest'
      responses:
        '200':
          description: Invite person after the assignment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '401':
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Combined permissions do not include the required manage_* permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Invite or role not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The invite already has this role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: duplicate_assignment
                message: That role is already assigned.
  /people/invites/{inviteId}/roles/{roleId}:
    delete:
      operationId: revokeInviteRole
      summary: Revoke one role from an unused invite
      description: 'Removes one InviteRole. Remaining roles stay. Last-admin does not apply (unused invites
        cannot administer). Requires manage_users.

        '
      tags:
      - Settings
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/InviteId'
      - $ref: '#/components/parameters/RoleId'
      responses:
        '200':
          description: Invite person after the revoke.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '401':
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Combined permissions do not include the required manage_* permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Invite, role, or assignment not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: FlyAccounts access JWT from POST /auth/microsoft or POST /auth/refresh.
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RoleId:
      name: roleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    InviteId:
      name: inviteId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ListSearch:
      name: search
      in: query
      required: false
      description: Case-insensitive search across relevant text fields.
      schema:
        type: string
    ListPage:
      name: page
      in: query
      required: false
      description: 1-based page index.
      schema:
        type: integer
        minimum: 1
        default: 1
    ListPageSize:
      name: pageSize
      in: query
      required: false
      description: Page size.
      schema:
        type: integer
        enum:
        - 10
        - 25
        - 50
        default: 10
...
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
      - code
      - message
      properties:
        code:
          type: string
          enum:
          - invalid_token
          - personal_account
          - unknown_tenant
          - unauthorized
          - forbidden
          - pending_access
          - not_found
          - duplicate_assignment
          - last_admin_required
          - duplicate_invite
          - already_present
          - duplicate_role_name
          - role_still_assigned
          - duplicate_permission
          - validation_error
          - entity_context_required
          - invalid_currency
          - invalid_file_type
          - client_in_use
          - duplicate_client_name
        message:
          type: string
...
    RoleSummary:
      type: object
      additionalProperties: false
      required:
      - id
      - name
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
...
    Person:
      type: object
      additionalProperties: false
      required:
      - id
      - personType
      - email
      - status
      - roles
      properties:
        id:
          type: string
          format: uuid
          description: User.id when personType is user; Invite.id when invite.
        personType:
          type: string
          enum:
          - user
          - invite
        email:
          type: string
          description: User.upn or Invite.email.
        displayName:
          type: string
          description: Present after sign-in; omitted for unused invites.
        status:
          type: string
          enum:
          - invited
          - pending
          - active
        entryPath:
          type: string
          enum:
          - invite
          - organization
          description: How a signed-in User entered. Omitted for unused invites.
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleSummary'
    PeopleResponse:
      type: object
      additionalProperties: false
      required:
      - people
      - page
      - pageSize
      - total
      properties:
        people:
          type: array
          items:
            $ref: '#/components/schemas/Person'
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          enum:
          - 10
          - 25
          - 50
        total:
          type: integer
          minimum: 0
...
    AssignRoleRequest:
      type: object
      additionalProperties: false
      required:
      - roleIds
      properties:
        roleIds:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
    CreateInviteRequest:
      type: object
      additionalProperties: false
      required:
      - email
      properties:
        email:
          type: string
          description: Organizational work email. Matching on sign-in is case-insensitive.
        roleIds:
          type: array
          items:
            type: string
            format: uuid
          default: []
          description: Existing roles to pre-assign. Empty means invite with no roles.
...